% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berg_distribution.R
\name{berg_sk_kt}
\alias{berg_sk_kt}
\alias{skberg}
\alias{ktberg}
\title{The BerG Skewness and Kurtosis}
\usage{
skberg(mu, phi)

ktberg(mu, phi)
}
\arguments{
\item{mu}{numeric; non-negative mean.}

\item{phi}{numeric; the dispersion index (greather than \code{abs(mu - 1)}).}
}
\value{
The functions \code{skberg} and \code{ktberg} returns the skewness and kurtosis of the BerG distribution
respectively.
}
\description{
Skewness and kurtosis for the BerG distribution
with parameters \code{mu} and \code{phi}.
}
\examples{
# For mu = 2 and phi = 2.5, the skewness and kurtosis of the BerG
# distribution are, respectively,
skberg(mu = 2, phi = 2.5)
ktberg(mu = 2, phi = 2.5)
}
\references{
Bourguignon, M. & Medeiros, R. (2020). A simple and useful regression model for fitting count data
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
