% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berg_methods.R
\name{bergrm-methods}
\alias{bergrm-methods}
\alias{plot.bergrm}
\alias{logLik.bergrm}
\alias{AIC.bergrm}
\alias{BIC.bergrm}
\alias{coef.bergrm}
\alias{vcov.bergrm}
\alias{model.matrix.bergrm}
\title{Methods for 'bergrm' objects}
\usage{
\method{plot}{bergrm}(object)

\method{logLik}{bergrm}(object)

\method{AIC}{bergrm}(object)

\method{BIC}{bergrm}(object)

\method{coef}{bergrm}(object, what = c("all", "mean", "dispersion"))

\method{vcov}{bergrm}(object)

\method{model.matrix}{bergrm}(object)
}
\arguments{
\item{object}{an object of class \code{bergrm}.}

\item{what}{a character indicating which parameter coefficients are
required, parameters for the \code{"mean"} or for the
\code{"dispersion"} model. If \code{"all"} (default), a list with
coefficients for the \code{mean} and for the \code{dispersion}
model is returned.}
}
\value{
.
}
\description{
Methods for 'bergrm' objects
}
\references{
Bourguignon, M. & Medeiros, R. (2019). A flexibe, simple and useful regression model for fitting count data.
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
