% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berg_estimation.R
\name{bergrm_estimation}
\alias{bergrm_estimation}
\alias{mle_berg}
\title{Maximum likelihood estimates for the BerG regression model}
\usage{
mle_berg(
  y,
  X,
  Z,
  link = "log",
  link.phi = "log",
  control = berg_control(...),
  eq_constraint = NULL,
  eq_constraint_jac = NULL,
  ...
)
}
\arguments{
\item{y}{a numeric vector of the response variable, in counts.}

\item{X, Z}{model matrices associated with the mean and the dispersion index, respectively.}

\item{link, link.phi}{character specification of the link function in the mean and in the dispersion index. The links \code{"log"} (default) \code{"sqrt"} and \code{"identity"} can be used.}

\item{control}{a list of control arguments specified via \code{berg_control}.}

\item{eq_constraint}{a function to evaluate equality constraints under the null hypothesis, in hypothesis testing contexts.}

\item{eq_constraint_jac}{a function to evaluate the jacobian of the equality constraints, which were passed by the \code{eq_constraint} argument.}

\item{...}{arguments passed to \code{berg_control}.}
}
\value{
A list with the values of the maximum likelihood estimates (\code{est});
the value of the log likelihood at estimates (\code{logLik}); and a logical value,
where \code{TRUE} indicates that the point obtained is feasible (\code{feasible}).
}
\description{
Maximum likelihood estimates of the coefficients of a fit
of the BerG regression model.
}
\references{
Bourguignon, M. & Medeiros, R. (2020). A simple and useful regression model for fitting count data.
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
