% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bids}
\alias{bids}
\title{Takeover bids data}
\format{
A data frame with 126 observations on the following 12 variables.

\itemize{
\item \code{numbids}: number of bids recevied after the initial bid.
\item \code{leglrest}: indicator variable for legal defence by lawsuit.
\item \code{realrest}: indicator variable for proposed changes in asset structure.
\item \code{finrest}: indicator variable for proposed change in ownership structure.
\item \code{whtknght}: indicator varible for management invitation for friendly third-party bid.
\item \code{bidprem}: bid price divided by price 14 working days before bid.
\item \code{insthold}: percentage of stock held by institutions.
\item \code{size}: total book value of assets in billions of dollars.
\item \code{regulatn}: indicator variable for Department of Justice intervention.
}
}
\usage{
data(bids)
}
\description{
Data of the number of bids received by 126 U.S.
firms that were targets of tender offers from 1978 to 1985, along with some
explanatory variables on the defensive actions taken by management of target
firm, firm-specific characteristics and intervention taken by federal
regulators.
}
\details{
Data originally from Jaggia and Thosar (1993) and used
as an example in Cameron and Trivedi (2013) and Saez-Castillo
and Conde-Sanchez (2013).
}
\references{
Cameron, A.C., Trivedi, P.K. (2013). Regression Analysis of Count Data. Cambridge University Press, second edition. \cr

Jaggia, S., Thosar, S. (1993). Multiple Bids as a Consequence of Target Management Resistance. Review of Quantitative Finance and Accounting, 447-457. \cr

Saez-Castillo, A.J., Conde-Sanchez, A. (2013). A hyper-Poisson regression model for overdispersed and underdispersed count data. Computational Statistics and Data Analysis, 61, 148-157.doi: 10.1016/j.csda.2012.12.009
}
\keyword{datasets}
