% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berg_diagnostic.R
\name{envel_berg}
\alias{envel_berg}
\title{Envelope Graph of the BerG Regression Model}
\usage{
envel_berg(
  object,
  residual = c("all", "quantile", "pearson"),
  R = 99,
  control = berg_control(...),
  ...
)
}
\arguments{
\item{object}{object of class 'bergrm'.}

\item{residual}{character; specifies which residual should be produced
in the normal probability plot. The available arguments are "all" (default)
for both randomized quantile and Pearson residuals; "quantile" for randomized
quantile residuals; and "pearson" for Pearson residuals.}

\item{R}{number of replicates.}

\item{control}{a list of control arguments specified via \code{berg_control}.}

\item{...}{arguments passed to \code{berg_control}.}
}
\value{
Returns the normal envelope simulated probability plot of
Pearson residuals and random quantile residuals, respectively.
}
\description{
Provides the normal envelope simulated
probability plot of Pearson's residuals, and of the randomized
quantile residuals resulting from the BerG regression model fit.
}
\references{
Bourguignon, M. and Medeiros, R. (2019). A flexibe, simple and useful regression model for fitting count data.
Kleiber, C., & Zeileis, A. (2016). Visualizing count data regressions using rootograms. The American Statistician, 70, 296-303
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
