% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berg_diagnostic.R
\name{expected_berg}
\alias{expected_berg}
\alias{expect_berg}
\title{Expected Frequencies by the BerG Regression Model}
\usage{
expect_berg(y, mu, phi)
}
\arguments{
\item{y}{vector of the observed values of the response variable.}

\item{mu}{vector of the fitted values.}

\item{phi}{vector of the fitted values for dispersion index.}
}
\value{
Expected frequencies after fitting the BerG regression model
to a dataset.
}
\description{
Provides the expected frequencies after fit of
the BerG regression model according to the definition given in
Kleiber and Zeileis (2016).
}
\examples{

# Sample size
n <- 100
# Covariates
X <- cbind(rep(1,n), runif(n, 0,1), runif(n,0,1))
Z <- cbind(rep(1,n), runif(n, 0,1), runif(n,0,1))

# Parameters and relations
beta <- c(1, 1.2, 0.2)
gama <- c(2, 1.5, 1.2)

mu <- exp(X\%*\%beta)
phi <- exp(Z\%*\%gama)

y <- rberg(n, mu, phi)

expect_berg(y, mu, phi)

}
\references{
Bourguignon, M. & Medeiros, R. (2020). A simple and useful regression model for fitting count data

Kleiber, C., & Zeileis, A. (2016). Visualizing count data regressions using rootograms. The American Statistician, 70, 296-303
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
