% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berg_diagnostic.R
\name{rootogram_berg}
\alias{rootogram_berg}
\alias{root_berg}
\title{Rootograms for the BerG Regression Model}
\usage{
root_berg(object)
}
\arguments{
\item{object}{object of class 'bergrm'.}
}
\value{
The rootogram for a fitted BerG regression.
}
\description{
Provides a rootograms that graphically compare
 (square roots) of empirical frequencies with fitted frequencies.
 For details see Kleiber and Zeileis (2016).
}
\examples{

data(grazing)

# BerG fit
fit <- glm.bg(birds ~ ., data = grazing)

# Rootogram
root_berg(fit)

}
\references{
Bourguignon, M. & Medeiros, R. (2020). A simple and useful regression model for fitting count data.

Kleiber, C., & Zeileis, A. (2016). Visualizing count data regressions using rootograms. The American Statistician, 70, 296-303
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
