% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berg_diagnostic.R
\name{rqr_berg}
\alias{rqr_berg}
\title{Randomized quantile residuals in the BerG regression model}
\usage{
rqr_berg(y, mu, phi)
}
\arguments{
\item{y}{vector; observed values of the response variable.}

\item{mu}{fitted values.}

\item{phi}{fitted values for dispersion index.}
}
\value{
Randomized quantile residuals resulting from the fit of
the BerG regression model.
}
\description{
Randomized quantile residuals resulting from the fit of
the BerG regression model.
}
\references{
Bourguignon, M. and Medeiros, R. (2019). A flexibe, simple and useful regression model for fitting count data.
Dunn, P. K., and Smyth, G. K. (1996). Randomized quantile residuals. Journal of Computational and Graphical Statistics, 5, 236-244.
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
