% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berg_methods.R
\name{bergrm-methods}
\alias{bergrm-methods}
\alias{plot.bergrm}
\alias{logLik.bergrm}
\alias{AIC.bergrm}
\alias{coef.bergrm}
\alias{vcov.bergrm}
\alias{model.matrix.bergrm}
\title{Methods for 'bergrm' objects}
\usage{
\method{plot}{bergrm}(x, residual = c("all", "quantile", "pearson"), ...)

\method{logLik}{bergrm}(object, ...)

\method{AIC}{bergrm}(object, ..., k = 2)

\method{coef}{bergrm}(object, what = c("all", "mean", "dispersion"), ...)

\method{vcov}{bergrm}(object, ...)

\method{model.matrix}{bergrm}(object, matrix = c("all", "mean", "dispersion"), ...)
}
\arguments{
\item{residual}{character; specifies which residual should be produced
in the summary plot. The available arguments are "all" for both
randomized quantile and Pearson residuals; "quantile" for randomized
quantile residuals; and "pearson" for Pearson residuals.}

\item{...}{further additional arguments.}

\item{object}{an object of class \code{bergrm}.}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is the classical AIC.}

\item{what}{a character indicating which parameter coefficients are
required, parameters for the \code{"mean"} or for the
\code{"dispersion"} model. If \code{"all"} (default), a list with
coefficients for the \code{mean} and for the \code{dispersion}
model is returned.}

\item{matrix}{a character indicating which model matrix is
required, the model matrix for the mean (\code{"mean"}) or for the
dispersion parameter (\code{"dispersion"}). If \code{"all"} (default), a list with
with both matrices are returned.}
}
\value{
.
}
\description{
Methods for 'bergrm' objects
}
\references{
Bourguignon, M. and Medeiros, R. M. R. (2020). A simple
    and useful regression model for fitting count data.
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
