% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berg_methods.R
\name{bergrm-methods}
\alias{bergrm-methods}
\alias{plot.bergrm}
\alias{logLik.bergrm}
\alias{AIC.bergrm}
\alias{BIC.bergrm}
\alias{coef.bergrm}
\alias{vcov.bergrm}
\alias{model.matrix.bergrm}
\title{Methods for 'bergrm' objects}
\usage{
\method{plot}{bergrm}(object)

\method{logLik}{bergrm}(object)

\method{AIC}{bergrm}(object, k = 2)

\method{BIC}{bergrm}(object)

\method{coef}{bergrm}(object, what = c("all", "mean", "dispersion"))

\method{vcov}{bergrm}(object)

\method{model.matrix}{bergrm}(object, matrix = c("all", "mean", "dispersion"))
}
\arguments{
\item{object}{an object of class \code{bergrm}.}

\item{what}{a character indicating which parameter coefficients are
required, parameters for the \code{"mean"} or for the
\code{"dispersion"} model. If \code{"all"} (default), a list with
coefficients for the \code{mean} and for the \code{dispersion}
model is returned.}

\item{matrix}{a character indicating which model matrix is
required, the model matrix for the mean (\code{"mean"}) or for the
dispersion parameter (\code{"dispersion"}). If \code{"all"} (default), a list with
with both matrices are returned.}

\item{numeric, }{the penalty per parameter to be used; the default k = 2 is the classical AIC.}
}
\value{
.
}
\description{
Methods for 'bergrm' objects
}
\references{
Bourguignon, M. and Medeiros, R. M. R. (2020). A simple
    and useful regression model for fitting count data.
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
