% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berg_regression.R
\name{bergrm}
\alias{bergrm}
\alias{glm.bg}
\alias{berg_fit}
\title{BerG Regression for Count Data}
\usage{
glm.bg(
  formula,
  data,
  link = c("log", "sqrt", "identity"),
  link.phi = NULL,
  disp.test = FALSE,
  control = berg_control(...),
  ...
)

berg_fit(
  y,
  X,
  Z = NULL,
  link = "log",
  link.phi = "log",
  control = berg_control(...),
  ...
)
}
\arguments{
\item{formula}{description of the model, of the type \code{y ~ x} for covariates in mean only, or \code{y ~ x | z} to enter covariates in the dispersion index.}

\item{data}{an optional data frame containing the variables in the formula. By default the variables are taken from environment(formula).}

\item{link, link.phi}{character; specification of the link function in the mean and in the dispersion index. The links \code{"log"} (default) \code{"sqrt"} and \code{"identity"} can be used.}

\item{disp.test}{logical, if TRUE, the function \code{bergrm} returns the test for constant dispersion.}

\item{control}{a list of control arguments specified via \code{berg_control}.}

\item{...}{arguments passed to \code{berg_control}}

\item{y}{a numeric vector of the response variable, in counts. This argument is used in the \code{berg_fit} function to get the estimates and required quantities that will return in the object resulting from the \code{bergrm} function.}

\item{X, Z}{model matrices associated with the mean and the dispersion parameter, respectively, which are used in the \code{fit_sdl} function.}
}
\value{
The \code{bergrm} function returns an object of class "bergrm", which consists of a list with the following components.
\describe{
  \item{coefficients}{ a list containing the elements "mean" and
        "dispersion" that consist of the estimates of the
         coefficients associated with the mean and the dispersion index, respectively.}
  \item{link, link.phi}{the link function used for the mean model, and for the dispersion index model, respectively.}
  \item{vcov}{asymptotic covariance matrix of the maximum likelihood estimator of the model parameters vector.}
  \item{logLik}{log-likelihood of the fitted model.}
  \item{AIC}{model Akaike information criteria.}
  \item{BIC}{model bayesian information criteria.}
  \item{n.obs, p, k}{Sample size, number of coefficients in the mean model, and number of coefficients in the dispersion index model, respectively.}
  \item{feasible}{Logical. If \code{TRUE}, the estimates obtained belong to the parametric space.}
  \item{pearson.residuals}{a vector with the Pearson residuals.}
  \item{residuals}{a vector with the randomized quantile residuals.}
  \item{fitted.values}{a vector with the fitted means.}
  \item{phi.hat}{a vector with the fitted dispersion indexes.}
  \item{linear.predictors}{a list containing the elements "eta1" and
        "eta2" that consist of the fitted linear predictor for the mean and the dispersion index.}
  \item{response}{the vector of the response.}
  \item{X, Z}{model matrices associated with the mean and the dispersion parameter, respectively.}
  \item{call}{the function call.}
  \item{formula}{the formula used to specify the model in \code{bergrm}.}
 }
}
\description{
Fit of the BerG regression model via maximum
likelihood for a new parameterization of this distribution
that is indexed by the mean and the dispersion index.
}
\references{
Bourguignon, M. & Medeiros, R. (2020). A simple and useful regression model for fitting count data
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
