% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bids}
\alias{bids}
\title{Takeover bids data}
\format{A data frame with 126 observations on the following 12 variables.

\itemize{
\item \code{docno}: a vector.
\item \code{weeks}: a vector.
\item \code{numbids}: a vector.
\item \code{bidprem}: a vector.
\item \code{insthold}: a vector.
\item \code{size}: a vector.
\item \code{leglrest}: a vector.
\item \code{realrest}: a vector.
\item \code{finrest}: a vector.
\item \code{regulatn}: a vector.
\item \code{whtknght}: a vector.
\item \code{sizeq}: a vector.
}}
\usage{
data(bids)
}
\description{
Data of the number of bids received by 126 U.S.
firms that were targets of tender offers from 1978 to 1985.
}
\details{
Data originally from Jaggia and Thosar (1993) and used
as an example in Cameron and Trivedi (2013) and Saez-Castillo
and Conde-Sanchez (2013).
}
\references{
Cameron, A.C., Trivedi, P.K. (2013). Regression Analysis of Count Data. Cambridge University Press, second edition. \cr

Jaggia, S., Thosar, S. (1993). Multiple Bids as a Consequence of Target Management Resistance. Review of Quantitative Finance and Accounting, 447-457. \cr

Saez-Castillo, A.J., Conde-Sanchez, A. (2013). A hyper-Poisson regression model for overdispersed and underdispersed count data. Computational Statistics and Data Analysis, 61, 148-157.doi: 10.1016/j.csda.2012.12.009
}
\keyword{datasets}
